requirements.txt
alabaster==0.7.12
astroid==2.9.0
Babel==2.9.1
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
dnspython==2.1.0
docopt==0.6.2
docutils==0.17.1
fett==0.3.2
greenlet==1.1.1
idna==3.3
imagesize==1.3.0
isort==5.10.1
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.6.0
livereload==2.6.3
MarkupSafe==2.0.1
mccabe==0.6.1
networkx==2.6.3
numpy==1.21.5
packaging==21.3
pandas==1.3.5
platformdirs==2.4.0
psycopg2==2.9.1
Pygments==2.10.0
pylint==2.12.2
pymongo==3.12.3
pyparsing==3.0.6
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jsonrpc-server==0.3.4
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
rstcheck==3.3.1
s3transfer==0.5.0
six==1.16.0
smart-open==5.2.1
snooty-lextudio==1.11.5
snowballstemmer==2.2.0
Sphinx==4.3.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
SQLAlchemy==1.4.23
sqlalchemy-cockroachdb==1.4.0
toml==0.10.2
tornado==6.1
typing_extensions==4.0.1
ujson==1.35
urllib3==1.26.7
watchdog==1.0.2
wrapt==1.13.3
